removed the possibly uneeded cast on values.gdbstub_port

+++ src/core/settings.cpp
@@ -15,7 +15,7 @@ Values values = {};
 
 void Apply() {
 
-    GDBStub::SetServerPort(static_cast<u32>(values.gdbstub_port));
+    GDBStub::SetServerPort(values.gdbstub_port);
     GDBStub::ToggleServer(values.use_gdbstub);
 
     VideoCore::g_hw_renderer_enabled = values.use_hw_renderer;
